backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pydantic[email]==2.5.0
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
cloudinary==1.36.0
